reduce number of times final output is compiled when loading a new example
The problem is that there is a race condition between inputs updated inside the module and inputs updated via the usual update<element>(). The module wins and triggers a render step, then the updates make it back to the server and trigger another render step. This commit primarily adds a flag to short-circuit compilation when loading the examples